requirements.txt
# Depth Anything V2 Fine-tuning Requirements

# PyTorch (CUDA 11.8 또는 12.1 버전에 맞게 설치)
torch>=2.0.0
torchvision>=0.15.0

# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
PyYAML>=6.0

# Training utilities
tensorboard>=2.14.0
tqdm>=4.66.0

# Hugging Face (DINOv2 encoder)
huggingface-hub>=0.17.0
timm>=0.9.0

# Optional: Mixed precision training
# apex  # NVIDIA Apex (optional, for mixed precision)
